Use autocannon to test connection pipelining on close
